requirements.txt
aiohttp==3.8.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
cachetools==4.2.4
certifi==2023.7.22
chardet==3.0.4
charset-normalizer==3.3.1
Django==4.2.6
django-environ==0.11.2
frozenlist==1.4.0
google-api-core==1.34.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
google-cloud-api-keys==0.5.4
google-cloud-core==1.7.3
google-cloud-translate==2.0.1
googleapis-common-protos==1.61.0
grpcio==1.59.2
grpcio-status==1.48.2
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2023.1.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
multidict==6.0.4
oauthlib==3.2.2
proto-plus==1.22.3
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyJWT==2.8.0
pyparsing==3.1.1
python-http-client==3.3.7
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
sendgrid==6.10.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
twilio==8.10.0
typing_extensions==4.8.0
urllib3==2.0.7
yarl==1.9.2
